+++ my-app/src/Sidebar/Sidebar.js
@@ -1,8 +1,7 @@
 /* eslint-disable import/no-anonymous-default-export */
 import React from 'react';
 import { slide as Menu } from 'react-burger-menu';
-import InicioTabContent from './Sidebar/InicioTabContent.js';
-import TabButtons from './Sidebar/TabButtons.js';
+import TabButtons from './TabButtons.js';
 
 export default function (props) {
   const [contenidoAMostrar, setContenidoAMostrar] = React.useState(null)
